+++ src/components/Dashboard.jsx
@@ -5,8 +5,9 @@ import UserEvents from "./UserEvents";
 import AuthorizedWrap from "./AuthorizedWrap";
 import AnalyzerWrap from "./AnalyzerWrap";
 import DashboardResults from "./DashboardResults";
-import {selectedEventNotSuccessful, noJobsFound} from "../app.messages";
+import {noJobsFound} from "../app.messages";
 import {connect} from "react-redux";
+import {faqUrl} from "../public/config";
 
 class Dashboard extends Component {
 
@@ -22,7 +23,18 @@ class Dashboard extends Component {
 			dashboardOutput = <DashboardResults />;
 		}
 		else if (this.props.isSelectedEventSuccessful === "error"){
-			dashboardOutput = <p className="error-message">{selectedEventNotSuccessful}</p>;
+			dashboardOutput = (
+					<div style={{display: "flex", justifyContent: "center"}}>
+						<p className="error-message">
+						The selected job did not complete successfully. The most common cause is overlapping
+						dates between the cash crop and the cover crop. Please check the field in 'My Farm' to
+						verify that the cover crop planting date does not overlap the cash crop harvest date and
+						the cover crop termination date does not overlap the next cash crop planting date.
+						For additional help, please see
+						the <a href={faqUrl} target="_blank" className="cc-link">FAQ</a> section.
+						</p>
+					</div>
+			);
 		}
 		else if (this.props.isSelectedEventSuccessful === "noRuns"){
 			dashboardOutput = <p className="error-message">{noJobsFound}</p>;