+++ src/public/config.js
@@ -1,3 +1,4 @@
 export const privacyUrl = "https://www.vpaa.uillinois.edu/resources/web_privacy";
 export const faqUrl = "https://opensource.ncsa.illinois.edu/confluence/display/CCrop/Frequently+Asked+Questions";
 export const croplandUrl = "https://nass.usda.gov/Research_and_Science/Cropland/SARS1a.php";
+export const nrecUrl = "https://www.illinoisnrec.org";